src/requirements.txt
ansicon==1.89.0
blessed==1.18.0
blessings==1.7
commonmark==0.9.1
jinxed==1.2.0
mariadb==1.1.8
mysql-connector-python==8.0.29
packaging==23.3
protobuf==3.20.1
pycryptodome==3.14.1
Pygments==2.12.0
pyperclip==1.8.2
rich==12.3.0
six==1.16.0
wcwidth==0.2.9
windows-curses==2.3.2
